Fixed last rect only receiving the update

+++ src/graphics/graphics.cpp
@@ -53,12 +53,12 @@ namespace SplitGui {
         IVec2 windowSize = pWindow->getSize();
 
         Vec2 newX1;
-        newX1.x = (float)x1.x / windowSize.x * 2.0 - 1.0f;
-        newX1.y = (float)x1.y / windowSize.y * 2.0 - 1.0f;
+        newX1.x = (float)x1.x / (float)windowSize.x * 2.0f - 1.0f;
+        newX1.y = (float)x1.y / (float)windowSize.y * 2.0f - 1.0f;
 
         Vec2 newX2;
-        newX2.x = (float)x2.x / windowSize.x * 2.0 - 1.0f;
-        newX2.y = (float)x2.y / windowSize.y * 2.0 - 1.0f;
+        newX2.x = (float)x2.x / (float)windowSize.x * 2.0f - 1.0f;
+        newX2.y = (float)x2.y / (float)windowSize.y * 2.0f - 1.0f;
 
         pInterface->updateRect(ref, newX1, newX2);
     }
